Finding your Golf Score.
In the game of golf, each hole has a par, meaning, the average number of strokes a golfer is expected to make in order to sink the ball in the hole to complete the play. Depending on how far above or below par your strokes are, there is a different nickname.
This code takes in your par and strokes and provides you with your score for each hole.